requirements.txt
torch
torchaudio
streamlit
pyannote.audio==3.1.1
numpy
soundfile
python-dotenv
tqdm
